Separate request managers
