+++ backend/requestmanagers/requestmanager.py
@@ -1,0 +1,27 @@
+from flask import Flask
+import json
+
+
+class RequestManager(object):
+    def __init__(self):
+        self.get_counter = 0
+        self.post_counter = 0
+
+    def _respond(self, status_code: int = 200, body: dict | list = None, mimetype: str = "application/json"):
+        return Flask.response_class(
+            status=status_code,
+            response=json.dumps(body),
+            mimetype=mimetype
+        )
+
+    def get_index(self):
+        """
+        Handle a GET request on the index page.
+        :return: A valid response for the index page
+        """
+        self.get_counter += 1
+        body = {
+            "counter": self.get_counter
+        }
+
+        return self._respond(status_code=200, body=body)
